Implement new issue_transitions stream
Previously, the issue's transitions were in the schema for issues, but
the data was never actually populated due to the `transitions` option
missing from the `expand` parameter to the search endpoint.

This change moves transitions out into their own stream. The stream
operates the same way the comments and changelog streams work - when
data is synced for the issues stream and the transitions stream is
selected, then the data will be output for that stream.