Feature/use metadata (#21)
* Refactor tap to use metadata (#16)

* wip

* transform records using metadata

* use updated singer-python to support patternProperties

* WIP: clarifying issues around envelope

* Initial commit

* WIP: Running spike, spike directory restructured

`tap-jira` was modified to make `looking_at_issues_and_fields.py`
as self-contained as possible. This was to try and minimize the
effect of code changes elsewhere on this spike.

Moreover, the restructuring was to address a clash in import
statements with regards to `http.py`. It appears that the
general name of `http.py` collided with an `http.py` file that
the `requests` library leans on.

* Create spike report

* Delete tap_config.json

* Remove commented code

* Remove unnecessary function on Context